added q2 plugin
