--- woltka/config.yml
@@ -1,3 +1,0 @@
---- # WoLtk configuration file
-
-...
